+++ src/main/java/generaloss/freetype/freetype/FTSizeMetrics.java
@@ -12,64 +12,64 @@ public class FTSizeMetrics extends FTStruct { // struct done.
 
 
     // FT_UShort x_ppem;
-    private static native int getXppem(long pointer);
+    private static native long getXppem(long pointer);
 
-    public int getXppem() {
+    public long getXppem() {
         return getXppem(super.pointer);
     }
 
     // FT_UShort y_ppem;
-    private static native int getYppem(long pointer);
+    private static native long getYppem(long pointer);
 
-    public int getYppem() {
+    public long getYppem() {
         return getYppem(super.pointer);
     }
 
     // FT_Fixed x_scale;
-    private static native int getXScale(long pointer);
+    private static native long getXScale(long pointer);
 
     public float getXScale() {
-        final int raw = getXScale(super.pointer);
+        final long raw = getXScale(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Fixed y_scale;
-    private static native int getYScale(long pointer);
+    private static native long getYScale(long pointer);
 
     public float getYScale() {
-        final int raw = getYScale(super.pointer);
+        final long raw = getYScale(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Pos ascender;
-    private static native int getAscender(long pointer);
+    private static native long getAscender(long pointer);
 
     public float getAscender() {
-        final int raw = getAscender(super.pointer);
+        final long raw = getAscender(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos descender;
-    private static native int getDescender(long pointer);
+    private static native long getDescender(long pointer);
 
     public float getDescender() {
-        final int raw = getDescender(super.pointer);
+        final long raw = getDescender(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos height;
-    private static native int getHeight(long pointer);
+    private static native long getHeight(long pointer);
 
     public float getHeight() {
-        final int raw = getHeight(super.pointer);
+        final long raw = getHeight(super.pointer);
         return FTPos.toFloat(raw);
     }
 
     // FT_Pos max_advance;
-    private static native int getMaxAdvance(long pointer);
+    private static native long getMaxAdvance(long pointer);
 
     public float getMaxAdvance() {
-        final int raw = getMaxAdvance(super.pointer);
+        final long raw = getMaxAdvance(super.pointer);
         return FTPos.toFloat(raw);
     }
