+++ src/main/java/generaloss/freetype/types/FTFixed.java
@@ -9,38 +9,38 @@ public class FTFixed extends FTStruct { // struct done.
     }
 
 
-    private static native int getRawValue(long pointer);
+    private static native long getRawValue(long pointer);
 
-    public int getRawValue() {
+    public long getRawValue() {
         return getRawValue(super.pointer);
     }
 
 
     public float getFloat() {
-        final int raw = this.getRawValue();
+        final long raw = this.getRawValue();
         return toFloat(raw);
     }
 
     public int getIntFloor() {
-        final int raw = this.getRawValue();
+        final long raw = this.getRawValue();
         return toIntFloor(raw);
     }
 
     public int getIntRound() {
-        final int raw = this.getRawValue();
+        final long raw = this.getRawValue();
         return toIntRound(raw);
     }
 
     public int getIntCeil() {
-        final int raw = this.getRawValue();
+        final long raw = this.getRawValue();
         return toIntCeil(raw);
     }
 
 
     @Override
     public String toString() {
-        final int raw = this.getRawValue();
-        return "FTFixed{float=" + toFloat(raw) + ", raw=0x" + Integer.toHexString(raw) + "}";
+        final long raw = this.getRawValue();
+        return "FTFixed{float=" + toFloat(raw) + ", raw=0x" + Long.toHexString(raw) + "}";
     }
 
 
@@ -52,30 +52,30 @@ public class FTFixed extends FTStruct { // struct done.
 
 
     private static final int BITS = 16;
-    private static final int UNIT = (1 << BITS); // 65536
-    private static final int HALF_UNIT = (1 << (BITS - 1)); // 32768
-    private static final int MASK = (UNIT - 1); // 0xFFFF
+    private static final long UNIT = (1 << BITS); // 65536
+    private static final long HALF_UNIT = (1 << (BITS - 1)); // 32768
+    private static final long MASK = (UNIT - 1); // 0xFFFF
     private static final float UNIT_FRAC = (1F / UNIT);
 
-    public static float toFloat(int rawValue) {
+    public static float toFloat(long rawValue) {
         return (rawValue * UNIT_FRAC);
     }
 
-    public static int toIntFloor(int rawValue) {
-        return (rawValue >> BITS);
+    public static int toIntFloor(long rawValue) {
+        return (int) (rawValue >> BITS);
     }
 
-    public static int toIntRound(int rawValue) {
-        return ((rawValue + HALF_UNIT) >> BITS);
+    public static int toIntRound(long rawValue) {
+        return (int) ((rawValue + HALF_UNIT) >> BITS);
     }
 
-    public static int toIntCeil(int rawValue) {
-        return (((rawValue + MASK) & ~MASK) >> BITS);
+    public static int toIntCeil(long rawValue) {
+        return (int) (((rawValue + MASK) & ~MASK) >> BITS);
     }
 
 
-    public static int of(float value) {
-        return (int) (value * UNIT);
+    public static long of(float value) {
+        return (long) (value * UNIT);
     }
 
 }
