+++ src/main/java/generaloss/freetype/types/FTMatrix.java
@@ -11,34 +11,34 @@ public class FTMatrix extends FTStruct { // struct done.
 
 
     // FT_Fixed xx;
-    private static native int getXX(long pointer);
+    private static native long getXX(long pointer);
 
     public float getXX() {
-        final int raw = getXX(super.pointer);
+        final long raw = getXX(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Fixed xy;
-    private static native int getXY(long pointer);
+    private static native long getXY(long pointer);
 
     public float getXY() {
-        final int raw = getXY(super.pointer);
+        final long raw = getXY(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Fixed yx;
-    private static native int getYX(long pointer);
+    private static native long getYX(long pointer);
 
     public float getYX() {
-        final int raw = getYX(super.pointer);
+        final long raw = getYX(super.pointer);
         return FTFixed.toFloat(raw);
     }
 
     // FT_Fixed yy;
-    private static native int getYY(long pointer);
+    private static native long getYY(long pointer);
 
     public float getYY() {
-        final int raw = getYY(super.pointer);
+        final long raw = getYY(super.pointer);
         return FTFixed.toFloat(raw);
     }
